move pooling test from test_nn to test/nn/test_pooling
